+++ src/api/admin.py
@@ -1,7 +1,7 @@
   
 import os
 from flask_admin import Admin
-from .models import db, User
+from .models import db, User, Favorites, Reviews, Comments, Tags, Shows, Watches
 from flask_admin.contrib.sqla import ModelView
 
 def setup_admin(app):
@@ -12,6 +12,12 @@ def setup_admin(app):
     
     # Add your models here, for example this is how we add a the User model to the admin
     admin.add_view(ModelView(User, db.session))
+    admin.add_view(ModelView(Favorites, db.session))
+    admin.add_view(ModelView(Reviews, db.session))
+    admin.add_view(ModelView(Comments, db.session))
+    admin.add_view(ModelView(Tags, db.session))
+    admin.add_view(ModelView(Shows, db.session))
+    admin.add_view(ModelView(Watches, db.session))
 
     # You can duplicate that line to add mew models
     # admin.add_view(ModelView(YourModelName, db.session))
